chore: split up devRegistry file
