+++ packages/wrangler/src/dev-registry/types.ts
@@ -1,0 +1,18 @@
+export type WorkerRegistry = Record<string, WorkerDefinition>;
+
+export type WorkerEntrypointsDefinition = Record<
+	"default" | string,
+	{ host: string; port: number } | undefined
+>;
+
+export type WorkerDefinition = {
+	port: number | undefined;
+	protocol: "http" | "https" | undefined;
+	host: string | undefined;
+	mode: "local" | "remote";
+	headers?: Record<string, string>;
+	entrypointAddresses?: WorkerEntrypointsDefinition;
+	durableObjects: { name: string; className: string }[];
+	durableObjectsHost?: string;
+	durableObjectsPort?: number;
+};
